Add break to prevent log flooding and API abuse
When running into connectivity or authentication problems, getRunnersOrganizationFromGithub gets stucked in a loop flooding the stdout with logs and hitting non-stop the Github API.
With a break, if it runs into an error of any sort, it will continue with the next org. If errors occurs with next orgs, it will only fail once and finally after checking all orgs it will pause the specified refresh value.